added forum access requests, reconnection feature, encrypted messages, leave private group and some minor updates

+++ groupcommunications-db/src/main/java/eu/h2020/helios_social/modules/groupcommunications/db/database/DatabaseComponentImpl.java
@@ -1,7 +1,10 @@
 package eu.h2020.helios_social.modules.groupcommunications.db.database;
+import java.security.KeyPair;
 
+import eu.h2020.helios_social.modules.groupcommunications.api.forum.sharing.ForumAccessRequest;
 import eu.h2020.helios_social.modules.groupcommunications.api.group.GroupMember;
 import eu.h2020.helios_social.modules.groupcommunications.api.resourcediscovery.EntityType;
+
 import eu.h2020.helios_social.modules.groupcommunications_utils.crypto.SecretKey;
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.CommitAction;
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.CommitAction.Visitor;
@@ -9,6 +12,7 @@ import eu.h2020.helios_social.modules.groupcommunications_utils.db.ContactExists
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.ContextExistsException;
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.DatabaseComponent;
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.DbCallable;
+import eu.h2020.helios_social.modules.groupcommunications_utils.db.GroupAccessRequestExistsException;
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.GroupInvitationExistsException;
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.NoSuchGroupException;
 import eu.h2020.helios_social.modules.groupcommunications_utils.db.NoSuchHeliosEventException;
@@ -22,6 +26,7 @@ import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.Conta
 import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.ContextInvitationAddedEvent;
 import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.ContextInvitationRemovedEvent;
 import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.ContextRenamedEvent;
+import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.GroupAccessRequestAddedEvent;
 import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.GroupAddedEvent;
 import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.GroupInvitationAddedEvent;
 import eu.h2020.helios_social.modules.groupcommunications_utils.sync.event.GroupInvitationRemovedEvent;
@@ -69,7 +74,6 @@ import eu.h2020.helios_social.modules.groupcommunications.api.contact.PendingCon
 import eu.h2020.helios_social.modules.groupcommunications.api.context.DBContext;
 import eu.h2020.helios_social.modules.groupcommunications.api.privategroup.sharing.GroupInvitation;
 import eu.h2020.helios_social.modules.groupcommunications.api.profile.Profile;
-import jdk.internal.jline.internal.Log;
 
 import java.util.Collection;
 import java.util.Map;
@@ -311,6 +315,13 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
     }
 
     @Override
+    public void removeGroupMember(Transaction transaction, GroupMember groupMember) throws DbException {
+        if (transaction.isReadOnly()) throw new IllegalArgumentException();
+        T txn = unbox(transaction);
+        db.removeGroupMember(txn, groupMember);
+    }
+
+    @Override
     public Collection<GroupMember> getGroupMembers(Transaction transaction, String groupid)
             throws DbException {
         T txn = unbox(transaction);
@@ -371,8 +382,11 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
             throw new NoSuchContextException();
         if (db.containsPendingGroupInvitation(txn,
                                               groupInvitations.getContactId(),
-                                              groupInvitations.getGroupId()))
+                                              groupInvitations.getGroupId()) ||
+            db.containsGroupAccessRequest(txn,groupInvitations.getContactId(),
+                    groupInvitations.getGroupId()))
             throw new GroupInvitationExistsException(groupInvitations);
+
         db.addGroupInvitation(txn, groupInvitations);
         transaction.attach(new GroupInvitationAddedEvent(groupInvitations));
     }
@@ -798,6 +812,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
         T txn = unbox(transaction);
         if (!db.containsContext(txn, context.getId())) {
             db.addContext(txn, context);
+            logDuration(LOG, "Context added event attached", 1);
             transaction.attach(new ContextAddedEvent(context));
         }
     }
@@ -874,12 +889,6 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
         return db.getContexts(txn);
     }
 
-    @Override
-    public Collection<DBContext> getContextsWithoutPrivateNames(Transaction transaction)
-            throws DbException {
-        T txn = unbox(transaction);
-        return db.getContextsWithoutPrivateName(txn);
-    }
 
     @Override
     public DBContext getContext(Transaction transaction,
@@ -976,13 +985,7 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
 
     }
 
-    @Override
-    public void addContextPrivateNameFeature(Transaction transaction) throws DbException {
-        if (transaction.isReadOnly()) throw new IllegalArgumentException();
-        T txn = unbox(transaction);
-        db.addContextPrivateNameFeature(txn);
 
-    }
 
     @Override
     public void setContextName(Transaction transaction, String contextId, String name) throws DbException {
@@ -1020,6 +1023,8 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
     @Override
     public void removePendingContext(Transaction transaction,
                                      String pendingContextId) throws DbException {
+        LOG.info("removePendingContext: "+pendingContextId);
+
         if (transaction.isReadOnly()) throw new IllegalArgumentException();
         T txn = unbox(transaction);
         if (!db.containsPendingContext(txn, pendingContextId))
@@ -1047,6 +1052,8 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
         if (!db.containsPendingContextInvitation(txn, contactId,
                                                  pendingContextId))
             throw new NoSuchPendingContextException();
+        LOG.info("removeContextInvitation: "+ pendingContextId);
+
         db.removeContextInvitation(txn, contactId, pendingContextId);
         transaction.attach(new ContextInvitationRemovedEvent(contactId,
                                                              pendingContextId));
@@ -1057,6 +1064,9 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
                                       ContactId contactId, String pendingGroupId) throws DbException {
         if (transaction.isReadOnly()) throw new IllegalArgumentException();
         T txn = unbox(transaction);
+        if (!db.containsPendingGroupInvitation(txn, contactId, pendingGroupId)){
+            return;
+        }
         if (!db.containsPendingGroup(txn, pendingGroupId))
             throw new NoSuchPendingGroupException();
         db.removeGroupInvitation(txn, contactId, pendingGroupId);
@@ -1128,6 +1138,87 @@ class DatabaseComponentImpl<T> implements DatabaseComponent {
             db.mergeSettings(txn, s, namespace);
             transaction.attach(new SettingsUpdatedEvent(namespace, merged));
         }
+    }
+
+    @Override
+    public void addGroupAccessRequest(Transaction transaction, ForumAccessRequest forumAccessRequest) throws DbException {
+            if (transaction.isReadOnly()) throw new IllegalArgumentException();
+            T txn = unbox(transaction);
+            if (forumAccessRequest.isIncoming() &&
+                    !db.containsGroup(txn, forumAccessRequest.getGroupId()))
+                throw new NoSuchGroupException();
+            if (!db.containsContext(txn, forumAccessRequest.getContextId()))
+                throw new NoSuchContextException();
+            if (db.containsGroupAccessRequest(txn,
+                    forumAccessRequest.getContactId(),
+                    forumAccessRequest.getGroupId()))
+                throw new GroupAccessRequestExistsException(forumAccessRequest);
+            db.addGroupAccessRequest(txn, forumAccessRequest);
+            transaction.attach(new GroupAccessRequestAddedEvent(forumAccessRequest));
+    }
+
+
+    @Override
+    public Collection<ForumAccessRequest> getGroupAccessRequests(Transaction transaction) throws DbException {
+        T txn = unbox(transaction);
+        return db.getGroupAccessRequests(txn);
+    }
+
+
+    @Override
+    public void removeGroupAccessRequest(Transaction transaction, ContactId contactId, String pendingGroupId) throws DbException {
+        if (transaction.isReadOnly()) throw new IllegalArgumentException();
+        T txn = unbox(transaction);
+        if (!db.containsGroupAccessRequestByGroupId(txn, pendingGroupId))
+            throw new NoSuchPendingGroupException();
+        db.removeGroupAccessRequest(txn, contactId, pendingGroupId);
+/*        transaction.attach(new GroupAccessRequestRemovedEvent(contactId,
+                pendingGroupId));*/
+    }
+
+    @Override
+    public boolean containsGroupAccessRequestByGroupId(Transaction transaction, String pendingGroupId) throws DbException {
+        T txn = unbox(transaction);
+        return db.containsGroupAccessRequestByGroupId(txn, pendingGroupId);
+    }
+
+    @Override
+    public int countUnreadMessagesInContext(Transaction transaction, String contextId) throws DbException {
+        T txn = unbox(transaction);
+        return db.countUnreadMessagesInContext(txn, contextId);
+    }
+
+    @Override
+    public int countGroupAccessRequest(Transaction transaction,
+                                            boolean isIncoming) throws DbException {
+        T txn = unbox(transaction);
+        return db.countGroupAccessRequests(txn, isIncoming);
+    }
+
+    @Override
+    public void addCryptoKeys(Transaction transaction, KeyPair keyPair)
+            throws DbException {
+        if (transaction.isReadOnly()) throw new IllegalArgumentException();
+        T txn = unbox(transaction);
+        if (!db.containsCryptoKeyPair(txn)) {
+            db.addCryptoKeys(txn, keyPair);
+        }
+    }
+
+    @Override
+    public boolean containsCryptoKeys(Transaction transaction)
+            throws DbException {
+        T txn = unbox(transaction);
+        return db.containsCryptoKeyPair(txn);
+    }
+
+    @Override
+    public KeyPair getCryptoKeys(Transaction transaction)
+            throws DbException {
+        T txn = unbox(transaction);
+        if (!db.containsCryptoKeyPair(txn))
+            throw new NoSuchIdentityException();
+        return db.getCryptoKeys(txn);
     }
 
     private class CommitActionVisitor implements Visitor {
